requirements.txt
interactions.py
aio-mc-rcon
cryptography
coloredlogs
